Implment DNS workaround
When using rustls, it's not possible to connect to a HTTPS endpoint by
using its IP address. That's a known limitation of rustls.

To circumvent this limitation we do the following operations when the request
is targeting an API server by using its IP address:

* Rewrite the request to target a well known FQDN of Kubernetes:
  `kubernetes.default.svc`
* Instruct reqwest to resolve the `kubernetes.default.svc` domain to the
  original IP address used inside of the request

This is something krew-wasm can do because it has a really specific use
case. Because of that, the repo now includes a custom implementation of
the exporter of the WIT http-request definitions.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>